backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
werkzeug==3.0.1
